Refactor package structure and modify existing DTO classes
- Separate entry and match on the model package
- Changed the paths of existing TftStatusDto and TftLeagueEntryResponse for consistency
- Improved management of TFT match information with the addition of TftMatch related model classes
