+++ api/src/main/java/com/glennsyj/rivals/api/rivalry/model/ParticipantStatDto.java
@@ -1,6 +1,6 @@
 package com.glennsyj.rivals.api.rivalry.model;
 
-import com.glennsyj.rivals.api.tft.model.TftStatusDto;
+import com.glennsyj.rivals.api.tft.model.entry.TftStatusDto;
 
 /*
     추후 확장될 가능성 있음
